+++ src/config.rs
@@ -59,6 +59,8 @@ pub struct GeneralConfig {
     pub auto_update_check: bool,
     #[serde(default)]
     pub auto_update_install: bool,
+    #[serde(default = "default_true")]
+    pub shell_integration: bool,
 }
 
 impl Default for GeneralConfig {
@@ -69,6 +71,7 @@ impl Default for GeneralConfig {
             visual_bell: false,
             auto_update_check: default_true(),
             auto_update_install: false,
+            shell_integration: default_true(),
         }
     }
 }